fix bugs in calculation and dates
